// src/app/interfaces/personaForm.interface.ts
// tslint:disable-next-line: class-name
export interface personaForm {
        nombre: string;
        apellido: string;
        fechaNacimiento: string;
        documento: string;
        cuil: string;
        sexo: string;
        calle: string;
        numero: string;
        piso: string;
        departamento: string;
        cpostal: string;
        barrio: string;
        monoblock: string;
        zona: string;
        localidad: string;
        provincia: string;
        pais: string;
        img: string;
        origenf: string;
        usuario: string;
        _id?: string
}